+++ packages/react-components/react-migration-v0-v9/src/components/StyledText/index.tsx
@@ -1,0 +1,2 @@
+export * from './StyledText';
+export * from './StyledText.styles';