+++ packages/react-components/react-provider/src/components/FluentProvider/FluentProvider.types.ts
@@ -4,6 +4,7 @@ import type {
   TooltipVisibilityContextValue_unstable as TooltipVisibilityContextValue,
   ThemeClassNameContextValue_unstable as ThemeClassNameContextValue,
   ThemeContextValue_unstable as ThemeContextValue,
+  CustomStyleHooksContextValue_unstable as CustomStyleHooksContextValue,
 } from '@fluentui/react-shared-contexts';
 import type { PartialTheme } from '@fluentui/react-theme';
 import type { ComponentProps, ComponentState, Slot } from '@fluentui/react-utilities';
@@ -12,12 +13,20 @@ export type FluentProviderSlots = {
   root: Slot<'div'>;
 };
 
+// exported for callers to avoid referencing react-shared-context
+// and applying Partial<> when passing custom style hooks.
+export type FluentProviderCustomStyleHooks = Partial<CustomStyleHooksContextValue>;
+
 export type FluentProviderProps = Omit<ComponentProps<FluentProviderSlots>, 'dir'> & {
   /**
    * Passes styles applied to a component down to portals if enabled.
    * @default true
    */
   applyStylesToPortals?: boolean;
+
+  /** Sets the hooks for custom styling components. */
+  // eslint-disable-next-line @typescript-eslint/naming-convention
+  customStyleHooks_unstable?: FluentProviderCustomStyleHooks;
 
   /** Sets the direction of text & generated styles. */
   dir?: 'ltr' | 'rtl';
@@ -34,12 +43,17 @@ export type FluentProviderProps = Omit<ComponentProps<FluentProviderSlots>, 'dir
 
 export type FluentProviderState = ComponentState<FluentProviderSlots> &
   Pick<FluentProviderProps, 'targetDocument'> &
-  Required<Pick<FluentProviderProps, 'applyStylesToPortals' | 'dir' | 'overrides_unstable'>> & {
+  Required<
+    Pick<FluentProviderProps, 'applyStylesToPortals' | 'customStyleHooks_unstable' | 'dir' | 'overrides_unstable'>
+  > & {
     theme: ThemeContextValue;
     themeClassName: string;
   };
 
-export type FluentProviderContextValues = Pick<FluentProviderState, 'theme' | 'overrides_unstable'> & {
+export type FluentProviderContextValues = Pick<
+  FluentProviderState,
+  'customStyleHooks_unstable' | 'theme' | 'overrides_unstable'
+> & {
   provider: ProviderContextValue;
   themeClassName: ThemeClassNameContextValue;
   textDirection: 'ltr' | 'rtl';