+++ packages/react-components/react-tree/src/index.ts
@@ -6,10 +6,21 @@ export {
   useTree_unstable,
   useTreeContextValues_unstable,
 } from './Tree';
-export type { TreeProps, TreeState, TreeSlots, TreeOpenChangeData, TreeOpenChangeEvent, TreeItemId } from './Tree';
+export type {
+  TreeProps,
+  TreeState,
+  TreeSlots,
+  TreeOpenChangeData,
+  TreeOpenChangeEvent,
+  TreeNavigationEvent_unstable,
+  TreeNavigationData_unstable,
+  TreeItemId,
+} from './Tree';
 
 export { TreeProvider, useTreeContext_unstable, useTreeItemContext_unstable, TreeItemProvider } from './contexts';
 export type { TreeContextValue } from './contexts';
+
+export { treeItemLevelToken } from './utils/tokens';
 
 export {
   TreeItem,
@@ -43,7 +54,13 @@ export type {
 } from './TreeItemPersonaLayout';
 
 export { useFlatTreeItems_unstable } from './hooks/index';
-export type { FlatTreeItemProps, FlatTreeProps, UseFlatTreeItemsOptions, LazyFlatTreeItems } from './hooks/index';
+export type {
+  FlatTreeItem,
+  FlatTreeItemProps,
+  FlatTreeProps,
+  UseFlatTreeItemsOptions,
+  LazyFlatTreeItems,
+} from './hooks/index';
 
 export { flattenTree_unstable } from './utils/flattenTree';
 export type { NestedTreeItem } from './utils/flattenTree';