+++ packages/react-examples/src/react-experiments/Tile/Tile.Document.Example.tsx
@@ -41,7 +41,7 @@ const ITEMS: { name: string; activity: string }[] = [
 interface IDocumentTileWithThumbnailProps {
   originalImageSize: ISize;
   showForeground: boolean;
-  item: typeof ITEMS[0];
+  item: (typeof ITEMS)[0];
 }
 
 const DocumentTileWithThumbnail: React.FunctionComponent<IDocumentTileWithThumbnailProps> = (