+++ packages/react-examples/src/react/ExtendedPeoplePicker/ExtendedPeoplePicker.doc.tsx
@@ -4,8 +4,10 @@ import { ExtendedPeoplePickerControlledExample } from './ExtendedPeoplePicker.Co
 
 import { IDocPageProps } from '@fluentui/react/lib/common/DocPage.types';
 
-const ExtendedPeoplePickerBasicExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/ExtendedPeoplePicker.Basic.Example.tsx') as string;
-const ExtendedPeoplePickerControlledExampleCode = require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/ExtendedPeoplePicker.Controlled.Example.tsx') as string;
+const ExtendedPeoplePickerBasicExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/ExtendedPeoplePicker.Basic.Example.tsx') as string;
+const ExtendedPeoplePickerControlledExampleCode =
+  require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/ExtendedPeoplePicker.Controlled.Example.tsx') as string;
 
 export const ExtendedPeoplePickerPageProps: IDocPageProps = {
   title: 'ExtendedPeoplePicker',
@@ -24,10 +26,13 @@ export const ExtendedPeoplePickerPageProps: IDocPageProps = {
       view: <ExtendedPeoplePickerControlledExample />,
     },
   ],
-  overview: require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerOverview.md') as string,
-  bestPractices: require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerBestPractices.md') as string,
+  overview:
+    require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerOverview.md') as string,
+  bestPractices:
+    require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerBestPractices.md') as string,
   dos: require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerDos.md') as string,
-  donts: require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerDonts.md') as string,
+  donts:
+    require('!raw-loader?esModule=false!@fluentui/react-examples/src/react/ExtendedPeoplePicker/docs/ExtendedPeoplePickerDonts.md') as string,
   isHeaderVisible: true,
   isFeedbackVisible: true,
 };