+++ packages/react/src/components/ExtendedPicker/BaseExtendedPicker.tsx
@@ -19,7 +19,8 @@ export interface IBaseExtendedPickerState<T> {
 
 export class BaseExtendedPicker<T, P extends IBaseExtendedPickerProps<T>>
   extends React.Component<P, IBaseExtendedPickerState<T>>
-  implements IBaseExtendedPicker<T> {
+  implements IBaseExtendedPicker<T>
+{
   public floatingPicker = React.createRef<BaseFloatingPicker<T, IBaseFloatingPickerProps<T>>>();
   public selectedItemsList = React.createRef<BaseSelectedItemsList<T, IBaseSelectedItemsListProps<T>>>();