+++ packages/react/src/components/FloatingPicker/Suggestions/SuggestionsStore.ts
@@ -52,7 +52,7 @@ export class SuggestionsStore<T> {
         ariaLabel:
           this.getAriaLabel !== undefined
             ? this.getAriaLabel(suggestion)
-            : ((suggestion as any) as ITag).name ||
+            : (suggestion as any as ITag).name ||
               (<IPersonaProps>suggestion).text ||
               // eslint-disable-next-line deprecation/deprecation
               (<IPersonaProps>suggestion).primaryText,