+++ packages/test-utilities/src/index.ts
@@ -2,3 +2,4 @@ export { createTestContainer } from './createTestContainer';
 export { getCSSRules } from './getCSSRules';
 export { safeCreate } from './safeCreate';
 export { safeMount } from './safeMount';
+export { create } from './reactTestRenderer';